GM Feature: new games always contain a scene; deleting last scene creates new one; active scene is updated if deleted

+++ orm.py
@@ -103,13 +103,14 @@ class Game(db.Entity):
 		return self.getImageUrl(image_id)
 
 	def clear(self):
-		# remove all images
 		game_root = self.getImagePath()
-		for img in self.getAllImages():
-			path = os.path.join(game_root, img)
-			os.remove(path)
-		# remove game directory
-		os.rmdir(game_root)
+		if os.path.isdir(game_root):
+			# remove all images
+			for img in self.getAllImages():
+				path = os.path.join(game_root, img)
+				os.remove(path)
+			# remove game directory
+			os.rmdir(game_root)
 
 
 # --- UNIT TESTS --------------------------------------------------------------
